fix codacy issues
