+++ microservices-dashboard-server/src/main/java/be/ordina/msdashboard/security/strategies/DefaultApplier.java
@@ -35,5 +35,6 @@ public class DefaultApplier implements SecurityProtocolApplier {
      */
     @Override
     public void apply(HttpClientRequest<ByteBuf> request) {
+        // intentionally left empty
     }
 }
