+++ microservices-dashboard-server/src/test/java/be/ordina/msdashboard/security/strategies/MsDashboardServerJWTSecurityIntegrationTest.java
@@ -15,15 +15,15 @@
  *
  */
 
-package be.ordina.msdashboard.security;
+package be.ordina.msdashboard.security.strategies;
 
 import be.ordina.msdashboard.EnableMicroservicesDashboardServer;
 import be.ordina.msdashboard.MicroservicesDashboardServerApplicationTest;
+import be.ordina.msdashboard.config.JWTAuthenticationInitializerInterceptor;
 import be.ordina.msdashboard.security.filter.AuthHealthFilter;
 import be.ordina.msdashboard.security.filter.AuthIndexFilter;
 import be.ordina.msdashboard.security.filter.AuthMappingsFilter;
 import be.ordina.msdashboard.security.filter.AuthPactFilter;
-import be.ordina.msdashboard.security.strategies.StrategyFactory;
 import be.ordina.msdashboard.security.strategy.SecurityProtocol;
 import be.ordina.msdashboard.wiremock.InMemoryMockedConfiguration;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -156,18 +156,6 @@ public class MsDashboardServerJWTSecurityIntegrationTest {
         logger.info("BODY: " + body);
         JSONAssert.assertEquals(load("src/test/resources/MsDashboardServerJWTSecurityIntegrationTestEventsResponse.json"),
                 body, JSONCompareMode.LENIENT);
-    }
-
-    private void printLinks(Map<String, List> r) {
-        List<Object> nodes = (List<Object>) r.get(NODES);
-        List<Map<String, Integer>> links = (List<Map<String, Integer>>) r.get(LINKS);
-        for (Map<String, Integer> link : links) {
-            int sourceIndex = link.get("source");
-            int targetIndex = link.get("target");
-            String sourceNodeId = (String) ((Map) nodes.get(sourceIndex)).get(ID);
-            String targetNodeId = (String) ((Map) nodes.get(targetIndex)).get(ID);
-            logger.info("Graph contains links between: " + sourceNodeId + " and " + targetNodeId);
-        }
     }
 
     private static void assertLinkBetweenIds(Map<String, List> r, String source, String target) throws IOException {
